Toon Shading.
Well... It worked once, but didn't draw any of the toon shaded items. I
think the problem is it's calling the PrepareMaterial that has 4
parameters rather than the one with two, but I don't want to mess with
that one since that uses the Material info stuff and I don't want to
screw that up.
As far as I can tell though only the UI uses the material item stuff so
if we can get it to only call the 4 parameter function specifically for
the UI that might allow things to work.